Chore: Typo correction, rename var..

+++ microservices/review-service/src/main/java/microservices/core/review/services/ReviewServiceImpl.java
@@ -7,16 +7,17 @@ import microservices.core.review.domain.ReviewEntity;
 import microservices.core.review.repository.ReviewRepository;
 import microservices.util.exceptions.InvalidInputException;
 import microservices.util.http.ServiceUtil;
+import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Flux;
 import reactor.core.scheduler.Scheduler;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Supplier;
+import java.util.logging.Level;
 
 @RestController
 @RequiredArgsConstructor
@@ -38,7 +39,9 @@ public class ReviewServiceImpl implements ReviewService {
         
         if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
 
-        return asyncFlux(getByProductId(productId)).log();
+        LOG.info("Will get reviews for product with id={}", productId);
+
+        return asyncFlux(() -> Flux.fromIterable(getByProductId(productId))).log(null, Level.FINE);
     }
 
     protected List<ReviewDTO> getByProductId(int productId) {
@@ -53,12 +56,15 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public ReviewDTO createReview(ReviewDTO body) {
+        if (body.getProductId() < 1) throw new InvalidInputException("Invalid productId: " + body.getProductId());
+
         try {
             ReviewEntity reviewEntity = reviewMapper.DTOToEntity(body);
             ReviewEntity save = reviewRepository.save(reviewEntity);
 
             LOG.debug("createReview: created a review entity: {}/{}", body.getProductId(), body.getReviewId());
             return reviewMapper.entityToDTO(save);
+            
         } catch (DataIntegrityViolationException dive) {
             throw new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Review Id:" + body.getReviewId());
         }
@@ -70,11 +76,13 @@ public class ReviewServiceImpl implements ReviewService {
      */
     @Override
     public void deleteReviews(int productId) {
+        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
+
         LOG.debug("deleteReviews: tries to delete reviews for the product with productId: {}", productId);
         reviewRepository.deleteAll(reviewRepository.findByProductId(productId));
     }
 
-    private <T> Flux<T> asyncFlux(Iterable<T> iterable) {
-        return Flux.fromIterable(iterable).publishOn(scheduler);
+    private <T> Flux<T> asyncFlux(Supplier<Publisher<T>> publisherSupplier) {
+        return Flux.defer(publisherSupplier).subscribeOn(scheduler);
     }
 }
